pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-fraud-detection-app"
version = "0.1.0"
description = "ML Fraud Detection App"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.3",
    "uvicorn==0.32.0",
    "httpx==0.27.2",
    "asyncpg==0.30.0",
    "scikit-learn==1.6.1",
    "joblib==1.5.3",
    "pandas==2.3.3",
    "numpy>=2.0",
    "pydantic-settings==2.7.1",
    "logfire[fastapi]>=4.25.0",
    "ipython>=9.10.0",
    "python-multipart==0.0.20",
    "tortoise-orm==0.25.1",
    "scalar-fastapi>=1.8.0",
]

[tool.setuptools]
packages = ["api"]

[project.optional-dependencies]
dev = [
    "mypy>=1.15.0",
    "pyright>=1.1.394,<1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "coverage>=7.6.4",
    "pydantic-settings==2.7.1",
    "ruff>=0.9.6",
    "chainmock==0.8.1",
]

[dependency-groups]
dev = [
    "chainmock==0.8.1",
    "mypy>=1.15.0",
    "pyright>=1.1.394,<1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.6",
]

[tool.ruff.lint]
select = [
    "B",
    "BLE",
    "DTZ",
    "E",
    "EM",
    "F",
    "FA",
    "I",
    "PT",
    "RSE",
    "RUF",
    "S",
    "T20",
    "TRY",
    "UP",
    "W605",
]
ignore = [
    "E111",
    "E114",
    "E117",
    "E501",
    "PT011",
    "PT013",
    "S101",
    "S110",
    "S112",
    "TRY003",
    "UP007",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["DTZ", "S105"]
"main.py" = ["T20"]

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
filterwarnings = [
    "ignore:Please use `import python_multipart` instead\\.:PendingDeprecationWarning:fastapi\\.dependencies\\.utils",
]

[tool.mypy]
python_version = "3.11"

[[tool.mypy.overrides]]
module = ["joblib", "pandas"]
ignore_missing_imports = true
